Apply Jordan's review2

+++ src/main/java/com/klipwallet/membership/repository/OperatorRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.klipwallet.membership.entity.Operator;
 
-public interface OperatorRepository extends JpaRepository<Operator, Integer> {
+public interface OperatorRepository extends JpaRepository<Operator, Long> {
 }
